+++ kernel/sched/Makefile
@@ -2,7 +2,13 @@
 CFLAGS += -I .
 
 
-all: sched.o
+all: sched.o cfs.o
 
 sched.o: sched.c
-	$(CC) $(CFLAGS) -c sched.c -o sched.o
+	gcc $(CFLAGS) -c sched.c -o sched.o
+
+cfs.o: cfs.c
+	gcc $(CFLAGS) -c cfs.c -o cfs.o
+
+clean:
+	echo "Done."
