+++ kernel/sched/sched.c
@@ -2,7 +2,7 @@
 #include <common/kprint.h>
 #include <driver/video/video.h>
 #include <common/spinlock.h>
-
+#include <sched/cfs.h>
 
 struct sched_queue_t sched_cfs_ready_queue[MAX_CPU_NUM]; // 就绪队列
 
@@ -49,7 +49,7 @@ void sched_cfs_enqueue(struct process_control_block *pcb)
 
 /**
  * @brief 包裹shced_cfs_enqueue(),将PCB加入就绪队列
- * 
+ *
  * @param pcb
  */
 void sched_enqueue(struct process_control_block *pcb)
@@ -58,82 +58,12 @@ void sched_enqueue(struct process_control_block *pcb)
 }
 
 /**
- * @brief 调度函数
- *
- */
-void sched_cfs()
-{
-
-    cli();
-
-    current_pcb->flags &= ~PF_NEED_SCHED;
-    struct process_control_block *proc = sched_cfs_dequeue();
-    // kdebug("sched_cfs_ready_queue[proc_current_cpu_id].count = %d", sched_cfs_ready_queue[proc_current_cpu_id].count);
-    if (current_pcb->virtual_runtime >= proc->virtual_runtime || !(current_pcb->state & PROC_RUNNING)) // 当前进程运行时间大于了下一进程的运行时间，进行切换
-    {
-
-        if (current_pcb->state & PROC_RUNNING) // 本次切换由于时间片到期引发，则再次加入就绪队列，否则交由其它功能模块进行管理
-            sched_cfs_enqueue(current_pcb);
-        // kdebug("proc->pid=%d, count=%d", proc->pid, sched_cfs_ready_queue[proc_current_cpu_id].count);
-        if (sched_cfs_ready_queue[proc_current_cpu_id].cpu_exec_proc_jiffies <= 0)
-        {
-            switch (proc->priority)
-            {
-            case 0:
-            case 1:
-                sched_cfs_ready_queue[proc_current_cpu_id].cpu_exec_proc_jiffies = 4 / sched_cfs_ready_queue[proc_current_cpu_id].count;
-                break;
-            case 2:
-            default:
-                sched_cfs_ready_queue[proc_current_cpu_id].cpu_exec_proc_jiffies = (4 / sched_cfs_ready_queue[proc_current_cpu_id].count) << 2;
-                break;
-            }
-        }
-
-        // if (proc->pid == 0)
-        // {
-        //     kdebug("switch to pid0, current pid%ld, vrt=%ld      pid0 vrt=%ld", current_pcb->pid, current_pcb->virtual_runtime, proc->virtual_runtime);
-        //     if(current_pcb->state != PROC_RUNNING)
-        //         kdebug("current_pcb->state!=PROC_RUNNING");
-        // }
-        
-        process_switch_mm(proc);
-
-        switch_proc(current_pcb, proc);
-    }
-    else // 不进行切换
-    {
-        // kdebug("not switch.");
-        sched_cfs_enqueue(proc);
-
-        if (sched_cfs_ready_queue[proc_current_cpu_id].cpu_exec_proc_jiffies <= 0)
-        {
-            switch (proc->priority)
-            {
-            case 0:
-            case 1:
-                sched_cfs_ready_queue[proc_current_cpu_id].cpu_exec_proc_jiffies = 4 / sched_cfs_ready_queue[proc_current_cpu_id].count;
-                // sched_cfs_ready_queue.cpu_exec_proc_jiffies = 5;
-                break;
-            case 2:
-            default:
-                // sched_cfs_ready_queue.cpu_exec_proc_jiffies = 5;
-
-                sched_cfs_ready_queue[proc_current_cpu_id].cpu_exec_proc_jiffies = (4 / sched_cfs_ready_queue[proc_current_cpu_id].count) << 2;
-                break;
-            }
-        }
-    }
-
-    sti();
-}
-
-/**
  * @brief 包裹sched_cfs(),调度函数
- * 
+ *
  */
 void sched()
 {
+    kinfo("**************sched  Starting...");
     sched_cfs();
 }
 
@@ -143,7 +73,7 @@ void sched()
  */
 void sched_update_jiffies()
 {
-    
+
     switch (current_pcb->priority)
     {
     case 0:
@@ -168,6 +98,7 @@ void sched_update_jiffies()
  */
 void sched_init()
 {
+    kinfo("**************sched init Starting...");
     memset(&sched_cfs_ready_queue, 0, sizeof(struct sched_queue_t) * MAX_CPU_NUM);
     for (int i = 0; i < MAX_CPU_NUM; ++i)
     {
